corrected paths
